requirements.txt
BeautifulSoup4
pandas
selenium
request
playwright
psycopg2
sqlalchemy
#install tambien chromiom y tambein     install playwright  nos permiete ver coookies que sera de ayuda al monetp de realizar consutase  ne bakend